+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-09-29 18:25:02
+
+### Changed Files
+- `plans/fix-carrier-update-error-2025-01-29.md`
+- `server.js`
+- `src/services/base/localApi.ts`
+
+### Statistics
+```
+ 3 files changed, 257 insertions(+), 6 deletions(-)
+```
+
 ## [Checkpoint] - 2025-09-29 12:32:16
 
 ### Changed Files